Move DataFormStructureProps to DataFormStructure
